Crate som_core

This crate contains common types that are useful to be shared across multiple tools when manipulating SOM-related things.

Modules

ast

The SOM Abstract Syntax Tree definitions: the common parser output.

bytecode

The SOM bytecode definitions. Used only by the bytecode interpreter, so should maybe be moved.

cli_parser

For parsing command line arguments

core_classes

The SOM core classes.

interner

Facilities for string interning.