som_core

Module bytecode

The SOM bytecode definitions. Used only by the bytecode interpreter, so should maybe be moved.

Structs

Bytecode

Bytecode structure, mostly just used to have a Bytecode::* namespace for different bytecodes.

BytecodeIter

Enums

BcEntry

Constants

BC_SIZE_1_ARG

BC_SIZE_2_ARG

BC_SIZE_NO_ARGS

BC_SIZE_U16_ARG

Functions

read_u16

split_u16