som_core::bytecode

Struct Bytecode

pub struct Bytecode;

Bytecode structure, mostly just used to have a Bytecode::* namespace for different bytecodes.

Implementations

impl Bytecode

pub const DUP: u8 = 0u8

pub const PUSH_LOCAL: u8 = 1u8

pub const PUSH_NON_LOCAL: u8 = 2u8

pub const PUSH_ARG: u8 = 3u8

pub const PUSH_NON_LOCAL_ARG: u8 = 4u8

pub const PUSH_FIELD: u8 = 5u8

pub const PUSH_BLOCK: u8 = 6u8

pub const PUSH_CONSTANT: u8 = 7u8

pub const PUSH_GLOBAL: u8 = 8u8

pub const INC: u8 = 9u8

pub const DEC: u8 = 10u8

pub const PUSH_0: u8 = 11u8

pub const PUSH_1: u8 = 12u8

pub const PUSH_NIL: u8 = 13u8

pub const PUSH_SELF: u8 = 14u8

pub const POP: u8 = 15u8

pub const POP_LOCAL: u8 = 16u8

pub const POP_ARG: u8 = 17u8

pub const POP_FIELD: u8 = 18u8

pub const SEND_1: u8 = 19u8

pub const SEND_2: u8 = 20u8

pub const SEND_3: u8 = 21u8

pub const SEND_N: u8 = 22u8

pub const SUPER_SEND: u8 = 23u8

pub const RETURN_SELF: u8 = 24u8

pub const RETURN_LOCAL: u8 = 25u8

pub const RETURN_NON_LOCAL: u8 = 26u8

pub const JUMP: u8 = 27u8

pub const JUMP_BACKWARD: u8 = 28u8

pub const JUMP_ON_TRUE_TOP_NIL: u8 = 29u8

pub const JUMP_ON_FALSE_TOP_NIL: u8 = 30u8

pub const JUMP_ON_NIL_TOP_TOP: u8 = 31u8

pub const JUMP_ON_NOT_NIL_TOP_TOP: u8 = 32u8

pub const JUMP_ON_NIL_POP: u8 = 33u8

pub const JUMP_ON_NOT_NIL_POP: u8 = 34u8

pub const JUMP_ON_TRUE_POP: u8 = 35u8

pub const JUMP_ON_FALSE_POP: u8 = 36u8

pub const JUMP_IF_GREATER: u8 = 37u8

pub const DUP_2: u8 = 38u8

impl Bytecode

pub fn get_iter(bytecodes: &[u8]) -> BytecodeIter<'_>

pub fn padded_name(bc: u8) -> &'static str

Get the instruction’s name padded so that every padded names are of the same length.