som_core::core_classes

Struct CoreClasses

pub struct CoreClasses<ClassPtr> {

    pub object_class: ClassPtr,
    pub class_class: ClassPtr,
    pub metaclass_class: ClassPtr,
    pub nil_class: ClassPtr,
    pub integer_class: ClassPtr,
    pub double_class: ClassPtr,
    pub array_class: ClassPtr,
    pub method_class: ClassPtr,
    pub primitive_class: ClassPtr,
    pub symbol_class: ClassPtr,
    pub string_class: ClassPtr,
    pub system_class: ClassPtr,
    pub block_class: ClassPtr,
    pub block1_class: ClassPtr,
    pub block2_class: ClassPtr,
    pub block3_class: ClassPtr,
    pub boolean_class: ClassPtr,
    pub true_class: ClassPtr,
    pub false_class: ClassPtr,

}

The core classes of the SOM interpreter.

This struct allows to always keep a reference to important classes, even in case of modifications to global bindings by user-defined code.

Fields
object_class: ClassPtr

The Object class.

class_class: ClassPtr

The Class class.

metaclass_class: ClassPtr

The Class class.

nil_class: ClassPtr

The Nil class.

integer_class: ClassPtr

The Integer class.

double_class: ClassPtr

The Double class.

array_class: ClassPtr

The Array class.

method_class: ClassPtr

The Method class.

primitive_class: ClassPtr

The Primitive class.

symbol_class: ClassPtr

The Symbol class.

string_class: ClassPtr

The String class.

system_class: ClassPtr

The System class.

block_class: ClassPtr

The Block class.

block1_class: ClassPtr

The Block1 class.

block2_class: ClassPtr

The Block2 class.

block3_class: ClassPtr

The Block3 class.

boolean_class: ClassPtr

The Boolean class.

true_class: ClassPtr

The True class.

false_class: ClassPtr

The False class.

Implementations

impl<ClassPtr: Clone> CoreClasses<ClassPtr>

pub fn from_load_cls_fn<F>(load_system_cls: F) -> Self
where F: FnMut(&str, Option<&ClassPtr>) -> ClassPtr,

Loads core classes given a closure that returns a pointer to a class. TODO: also take a closure to set_class and set_super_class to do the rest of the hooking up.

impl<ClassPtr: Clone> CoreClasses<ClassPtr>

pub fn object_class(&self) -> ClassPtr

Get the Object class.

pub fn nil_class(&self) -> ClassPtr

Get the Nil class.

pub fn system_class(&self) -> ClassPtr

Get the System class.

pub fn symbol_class(&self) -> ClassPtr

Get the Symbol class.

pub fn string_class(&self) -> ClassPtr

Get the String class.

pub fn array_class(&self) -> ClassPtr

Get the Array class.

pub fn integer_class(&self) -> ClassPtr

Get the Integer class.

pub fn double_class(&self) -> ClassPtr

Get the Double class.

pub fn block_class(&self) -> ClassPtr

Get the Block class.

pub fn block1_class(&self) -> ClassPtr

Get the Block1 class.

pub fn block2_class(&self) -> ClassPtr

Get the Block2 class.

pub fn block3_class(&self) -> ClassPtr

Get the Block3 class.

pub fn true_class(&self) -> ClassPtr

Get the True class.

pub fn false_class(&self) -> ClassPtr

Get the False class.

pub fn metaclass_class(&self) -> ClassPtr

Get the Metaclass class.

pub fn method_class(&self) -> ClassPtr

Get the Method class.

pub fn primitive_class(&self) -> ClassPtr

Get the Primitive class.

impl<Class: Clone> CoreClasses<Class>

pub fn iter(&self) -> CoreClassesIter<'_, Class>